requirements.txt
ccxt
python-dotenv
telethon
requests
flask